Packer script updations

+++ packer/scripts/setup-webapp.sh
@@ -29,7 +29,12 @@ echo "Copying webapp & config from $webapp_zip_file to $webapp_dir and setting p
 
 mkdir -p "$webapp_dir"
 unzip -q -o "$webapp_zip_file" -d "$webapp_dir"
-cp -f "$webapp_config_file" "$webapp_dir"/dist/
+
+# Copy the configuration file if it exists
+if [ -f "$webapp_config_file" ]; then
+   cp -f "$webapp_config_file" "$webapp_dir"/dist/
+fi
+
 npm install --prefix "$webapp_dir"/dist/
 chown -R "$username":"$username" "$webapp_dir"
 chmod -R 755 "$webapp_dir"
